% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_targets.R
\name{build_targets_indices}
\alias{build_targets_indices}
\title{Build core targets for "indices" schema}
\usage{
build_targets_indices()
}
\value{
All the targets required, containing dynamic patterns.
}
\description{
Take advantage of the \code{\link[targets:tar_target_raw]{targets::tar_target_raw()}} function, and looping over
all the preprocessing functions.
}
