% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices.R
\name{calc_indices}
\alias{calc_indices}
\title{Calculate Indices}
\usage{
calc_indices(data, prep_fun, name_data = "game_data")
}
\arguments{
\item{data}{The raw data.}

\item{prep_fun}{The name (symbol) of the calculation function}

\item{name_data}{The column name in the \code{data} that stores original data. It
is typically of a vector containing \code{JSON} string.}
}
\value{
A \code{tibble} with the calculated indices.
}
\description{
Get raw data prepared and calculate indices.
}
\author{
Liang Zhang
}
