% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{check_source}
\alias{check_source}
\title{Check if the database based on the given data source is ready}
\usage{
check_source(source = setup_source())
}
\arguments{
\item{source}{The data source from which data is fetched. See
\code{\link[=setup_source]{setup_source()}} for details.}
}
\value{
TRUE if the database is ready, FALSE otherwise.
}
\description{
Check if the database based on the given data source is ready
}
