% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_user_info}
\alias{clean_user_info}
\title{Clean user information}
\usage{
clean_user_info(users, props)
}
\arguments{
\item{users}{A \link{data.frame} contains the user information.}

\item{props}{A character vector of the user properties to keep.}
}
\value{
A \link{data.frame} contains the cleaned user information.
}
\description{
Clean user information
}
