% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_users_props}
\alias{clean_users_props}
\title{Clean users properties}
\usage{
clean_users_props(users, props)
}
\arguments{
\item{users}{A \link{data.frame} contains the users properties.}

\item{props}{A character vector of the users properties to keep.}
}
\value{
A \link{data.frame} contains the cleaned users properties.
}
\description{
Clean users properties
}
