% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{config_file}
\alias{config_file}
\title{Configuration file}
\format{
An object of class \code{character} of length 1.
}
\usage{
config_file
}
\description{
Configuration file
}
\keyword{internal}
