% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_is_ready}
\alias{db_is_ready}
\title{Check if the database is ready}
\usage{
db_is_ready(source = setup_source())
}
\arguments{
\item{source}{The data source from which data is fetched. See
\code{\link[=setup_source]{setup_source()}} for details.}
}
\value{
TRUE if the database is ready, FALSE otherwise.
}
\description{
Check if the database is ready
}
