% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_config.R
\name{edit_config}
\alias{edit_config}
\title{Edit configuration file}
\usage{
edit_config(file = "config.yml")
}
\arguments{
\item{file}{The configuration file name.}
}
\value{
\code{NULL} (invisibly). This function is called for its side effect.
}
\description{
Open file "config.yml" to make it proper for your analysis. If not found,
this function will create one.
}
