% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\title{Fetch datasets from given database}
\usage{
fetch(query_file, config_where = NULL, dsn = "iquizoo-v3")
}
\arguments{
\item{query_file}{File name of \code{sql} query}

\item{config_where}{Configuration of "where-clause" of the \code{sql} query. Can
be a \code{list} (mostly from the \code{config.yml} file) or \code{data.frame}.}

\item{dsn}{Data source name of an \code{odbc} database connector.}
}
\value{
A \code{tibble} of original data
}
\description{
Based on a template query file and "where clause" configuration, datasets are
extracted from a given database.
}
\author{
Liang Zhang
}
