% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fetch_data.R
\name{fetch_batch}
\alias{fetch_batch}
\title{Fetch results of a parameterized query based on a batch of parameters}
\usage{
fetch_batch(query, params, ...)
}
\arguments{
\item{query}{A character string containing parameterized SQL.}

\item{params}{The parameters used in the SQL query.}

\item{...}{Further arguments passed to \code{\link[=fetch_parameterized]{fetch_parameterized()}}.}
}
\value{
A \link{data.frame} contains the fetched data.
}
\description{
Fetch results of a parameterized query based on a batch of parameters
}
