% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_from_v3.R
\name{fetch_from_v3}
\alias{fetch_from_v3}
\title{Fetch original data from database}
\usage{
fetch_from_v3(query_file, config_where = NULL)
}
\arguments{
\item{query_file}{File name of \code{sql} query}

\item{config_where}{Configuration of "where-clause" of the \code{sql} query. Can
be a \code{list} (mostly from the \code{config.yml} file) or \code{data.frame}.}
}
\value{
A \code{tibble} of original data
}
\description{
Based on a template query file, original data based on given "where clause"
configuration (\code{where_config}) are extracted from IQUIZOO database. If
\code{where_config} is \code{NULL}, this will extract all data.
}
\author{
Liang Zhang
}
