% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{fetch_iquizoo_mem}
\alias{fetch_iquizoo_mem}
\title{Memoised version of \code{\link[=fetch_iquizoo]{fetch_iquizoo()}}}
\usage{
fetch_iquizoo_mem(cache = NULL)
}
\arguments{
\item{cache}{The cache to be used. Default cache could be configured by
setting the environment variable \code{TARFLOW_CACHE} to \code{"disk"} or \code{"memory"}.
If set \code{TARFLOW_CACHE} to \code{"disk"}, the cache will be stored in disk at
\verb{~/.cache/tarflow.iquizoo}. If set \code{TARFLOW_CACHE} to \code{"memory"}, the cache
will be stored in memory. You can also set \code{cache} to a custom cache, see
\code{\link[memoise:memoise]{memoise::memoise()}} for more details.}
}
\value{
A memoised version of \code{\link[=fetch_iquizoo]{fetch_iquizoo()}}.
}
\description{
This function is a memoised version of \code{\link[=fetch_iquizoo]{fetch_iquizoo()}}. It is useful when
the same query is called multiple times or you want to cache the result. See
\code{\link[memoise:memoise]{memoise::memoise()}} and \code{\link[=fetch_iquizoo]{fetch_iquizoo()}} for more details.
}
\seealso{
\code{\link[=fetch_iquizoo]{fetch_iquizoo()}} for the original function.
}
