% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_parameterized.R
\name{fetch_parameterized}
\alias{fetch_parameterized}
\title{Fetch data from iQuizoo database based on a parameterized query}
\usage{
fetch_parameterized(query, params, ..., source = set_source())
}
\arguments{
\item{query}{A character string containing SQL.}

\item{params}{The parameters to be bound to the query. This parameter could
be safely omitted if \code{query} does not contain any parameters.}

\item{...}{Further arguments passed to \link[DBI:dbConnect]{DBI::dbConnect()}.}

\item{source}{The data source from which data is fetched. See \code{\link[=set_source]{set_source()}}
for details.}
}
\value{
A \link{data.frame} contains the fetched data.
}
\description{
Fetch data from iQuizoo database based on a parameterized query
}
