% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fetch_data.R
\name{fetch_preset}
\alias{fetch_preset}
\title{Fetch data from iQuizoo database with preset SQL query files}
\usage{
fetch_preset(..., what = c("project_contents", "project_users"))
}
\arguments{
\item{...}{The parameters used in the SQL query.}

\item{what}{The name of the preset query file to use.}
}
\value{
A \link{data.frame} contains the fetched data.
}
\description{
Fetch data from iQuizoo database with preset SQL query files
}
