% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_names.R
\name{get_game_names}
\alias{get_game_names}
\title{Find out the names of games to be processed}
\usage{
get_game_names(use_db = FALSE)
}
\arguments{
\item{use_db}{Use online database to get a better guess.}
}
\value{
The Chinese names of found games
}
\description{
The game names are mainly extracted from the \code{config.yml} file. Future work
might be extracted those games from given courses.
}
