% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_users_props_names}
\alias{get_users_props_names}
\title{Get the names of the user properties.}
\usage{
get_users_props_names()
}
\value{
A character vector of the names.
}
\description{
Get the names of the user properties.
}
