% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_where.R
\name{insert_where}
\alias{insert_where}
\alias{insert_where.NULL}
\alias{insert_where.list}
\alias{insert_where.data.frame}
\title{Insert where configurations}
\usage{
insert_where(old, ...)

\method{insert_where}{`NULL`}(old, ...)

\method{insert_where}{list}(old, element, ..., replace = TRUE)

\method{insert_where}{data.frame}(old, element, ..., replace = TRUE)
}
\arguments{
\item{old}{The old where configuration.}

\item{...}{Other argument kept for future use. Currently, they are silently
ignored.}

\item{element}{The new where element. Can be a \code{list} or \code{data.frame}. \code{NULL}
is accepted so that returns an unchanged version of the old config.}

\item{replace}{Replace the element with the same \code{table} value or not.
Default is set to replace, set \code{FALSE} to add it anyway (not recommended,
might cause error in further analysis).}
}
\value{
A where configuration of the same format with the old one.
}
\description{
Insert new where configurations into the existed one. This can be useful in
many projects doing minor changes to current where configurations.
}
\details{
This function is intended to modify an existed where configuration. To create
a completely new one, please specify it by using \code{\link[tibble:tribble]{tibble::tribble()}} instead.
}
\author{
Liang Zhang
}
