% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{parse_data}
\alias{parse_data}
\title{Parse Raw Data}
\usage{
parse_data(data, col_raw_json = "game_data", name_raw_parsed = "raw_parsed")
}
\arguments{
\item{data}{The raw data.}

\item{col_raw_json}{The column name storing raw json string data.}

\item{name_raw_parsed}{The name used to store parsed data.}
}
\value{
A \link{data.frame} contains the parsed data.
}
\description{
Raw data fetched from iQuizoo database is stored in json string format. This
function is used to parse raw json string data as \code{\link[=data.frame]{data.frame()}} and store
them in a list column.
}
