% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickup.R
\name{pickup}
\alias{pickup}
\title{Download (pick up) datasets from given database}
\usage{
pickup(
  query_file,
  config_where = NULL,
  dsn = "iquizoo-v3",
  groups = "iquizoo-v3",
  drv = getOption("tarflow.driver"),
  encoding = "utf-8"
)
}
\arguments{
\item{query_file}{File name of \code{sql} query. Literal query is acceptable, and
to be recognized as literal query, the input must be either wrapped with
\code{\link[=I]{I()}} or a string containing at least one new line.}

\item{config_where}{Configuration of "where-clause" of the \code{sql} query. Can
be a \code{list} (mostly from the \code{config.yml} file) or \code{data.frame}.}

\item{dsn}{The data source name of an \strong{ODBC} database connector. See
\code{\link[odbc:dbConnect-OdbcDriver-method]{odbc::dbConnect()}} for more information. Used when \code{drv} is set as
\code{\link[odbc:odbc]{odbc::odbc()}}.}

\item{groups}{Section identifier in the \code{default.file}. See
\code{\link[RMariaDB:dbConnect-MariaDBDriver-method]{RMariaDB::MariaDB()}} for more information. Used when \code{drv} is set as
\code{\link[RMariaDB:dbConnect-MariaDBDriver-method]{RMariaDB::MariaDB()}}.}

\item{drv}{The driver used. Set as an option of \code{"tarflow.driver"} and the
default is currently \code{odbc::odbc()}. Options are \code{\link[odbc:odbc]{odbc::odbc()}} and
\code{\link[RMariaDB:dbConnect-MariaDBDriver-method]{RMariaDB::MariaDB()}}, both of which need pre-configurations.}

\item{encoding}{Encoding to be assumed for input strings. Default to "UTF-8".}
}
\value{
A \link[tibble:tibble-package]{tibble} of the downloaded data.
}
\description{
Based on a template query file and "where clause" configuration, datasets are
extracted from a given database.
}
\author{
Liang Zhang
}
