% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fetch_data.R
\name{prepare_fetch_data}
\alias{prepare_fetch_data}
\title{Prepare targets pipeline for fetching data}
\usage{
prepare_fetch_data(params, ..., what = c("all", "raw_data", "scores"))
}
\arguments{
\item{params}{A \link{data.frame} contains the parameters to be bound to the
query. For now, only \code{organization_name} and \code{project_name} are supported
and both of them should be specified. Each row is a set of parameters.}

\item{...}{For future usage. Should be empty.}

\item{what}{What to fetch. Can be "all", "raw_data" or "scores".}
}
\value{
A S3 object of class \code{tarflow_targets}. The main component is a list
of targets. The other component is a \link{data.frame} contains the contents
based on which data is fetched.
}
\description{
Prepare targets pipeline for fetching data
}
