% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{preproc_data}
\alias{preproc_data}
\title{Feed Raw Data to Pre-processing}
\usage{
preproc_data(
  data,
  fn,
  ...,
  name_raw_parsed = "raw_parsed",
  pivot_results = TRUE,
  pivot_names_to = "index_name",
  pivot_values_to = "score"
)
}
\arguments{
\item{data}{A \link{data.frame} contains raw data.}

\item{fn}{This can be a function or formula. See \code{\link[rlang:as_function]{rlang::as_function()}} for
more details.}

\item{...}{Additional arguments passed to \code{fn}.}

\item{name_raw_parsed}{The column name in which stores user's raw data in
format of a list of \link{data.frame}s.}

\item{pivot_results}{Whether to pivot the calculated indices. If \code{TRUE}, the
calculated indices are pivoted into long format, with each index name
stored in the column of \code{pivot_names_to}, and each index value stored in
the column of \code{pivot_values_to}. If \code{FALSE}, the calculated indices are
stored in the same format as returned by \code{fn}.}

\item{pivot_names_to, pivot_values_to}{The column names used to store index
names and values if \code{pivot_results} is \code{TRUE}. See \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}
for more details.}
}
\value{
A \link{data.frame} contains the calculated indices.
}
\description{
Calculate indices using data typically returned by \code{\link[=wrangle_data]{wrangle_data()}}.
}
\details{
Observations with empty raw data (empty vector, e.g. \code{NULL}, in
\code{name_raw_parsed} column) are removed before calculating indices. If no
observations left after removing, a warning is signaled and \code{NULL} is
returned.
}
