% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{query_dir}
\alias{query_dir}
\title{Query directory name}
\format{
An object of class \code{character} of length 1.
}
\usage{
query_dir
}
\description{
Query directory name
}
\keyword{internal}
