% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{query_files}
\alias{query_files}
\title{Query files used}
\format{
An object of class \code{character} of length 5.
}
\usage{
query_files
}
\description{
Character vector. Name should be used as the name of the fetched object.
}
\keyword{internal}
