% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{schemas}
\alias{schemas}
\title{Supported schemas}
\format{
An object of class \code{character} of length 3.
}
\usage{
schemas
}
\description{
Character vector. Name is used as a short memo. Developers should describe it
in detail in the value because it is displayed in the wizard.
}
\keyword{internal}
