% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_games.R
\name{search_games}
\alias{search_games}
\title{Search games used in analysis}
\usage{
search_games(config_where)
}
\arguments{
\item{config_where}{Configuration of "where-clause".}
}
\value{
A \link[tibble:tibble-package]{tibble} contains all the games to be
analyzed and its related information.
}
\description{
When separate analysis into branches, the games to be analyzed should be
known in advance. This function should be used with memoise package to reduce
executing time.
}
\details{
The fetched will be joined with \code{dataproc.iquizoo::game_info} to expose the
pre-processing function name.
}
\examples{
\dontrun{
# use it with memoise
search_games_mem <- memoise::memoise(search_games)
games <- search_games_mem(config::get("where"))
}

}
