% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_parameterized.R
\name{set_source}
\alias{set_source}
\title{Set data source}
\usage{
set_source(
  driver = getOption("tarflow.driver"),
  dsn = getOption("tarflow.dsn"),
  groups = getOption("tarflow.groups")
)
}
\arguments{
\item{driver}{The driver used. Set as an option of \code{"tarflow.driver"}.
Options are \code{\link[odbc:odbc]{odbc::odbc()}} and \code{\link[RMariaDB:dbConnect-MariaDBDriver-method]{RMariaDB::MariaDB()}}, both of which need
pre-configurations. Default to first available one.}

\item{dsn}{The data source name of an \strong{ODBC} database connector. See
\code{\link[odbc:dbConnect-OdbcDriver-method]{odbc::dbConnect()}} for more information. Used when \code{driver} is set as
\code{\link[odbc:odbc]{odbc::odbc()}}.}

\item{groups}{Section identifier in the \code{default.file}. See
\code{\link[RMariaDB:dbConnect-MariaDBDriver-method]{RMariaDB::MariaDB()}} for more information. Used when \code{driver} is set as
\code{\link[RMariaDB:dbConnect-MariaDBDriver-method]{RMariaDB::MariaDB()}}.}
}
\value{
An S3 class of \code{tarflow.source} with the options.
}
\description{
Set data source
}
