% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{tar_fetch_data}
\alias{tar_fetch_data}
\title{Generate a set of targets for fetching data}
\usage{
tar_fetch_data(
  contents,
  templates = setup_templates(),
  what = c("raw_data", "scores"),
  check_progress = TRUE
)
}
\arguments{
\item{contents}{The contents structure used as the configuration of data
fetching.}

\item{templates}{The SQL template files used to fetch data. See
\code{\link[=setup_templates]{setup_templates()}} for details.}

\item{what}{What to fetch.}

\item{check_progress}{Whether to check the progress hash. If set as \code{TRUE},
Before fetching the data, the progress hash objects named as
\verb{progress_hash_\{project_id\}} will be depended on, which are typically
generated by \code{\link[=tar_prep_proj]{tar_prep_proj()}}. If the projects are finalized, set this
argument as \code{FALSE}.}
}
\value{
A list of target objects.
}
\description{
This target factory is the main part of the \code{tar_prep_iquizoo} function. It
fetches the raw data and scores for each project and task/game combination.
}
