% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{tar_fetch_users}
\alias{tar_fetch_users}
\title{Generate a set of targets for fetching user information}
\usage{
tar_fetch_users(contents, templates = setup_templates())
}
\arguments{
\item{contents}{The contents structure used as the configuration of data
fetching.}

\item{templates}{The SQL template files used to fetch data. See
\code{\link[=setup_templates]{setup_templates()}} for details.}
}
\value{
A list of target objects.
}
\description{
The user information is used to identify the users involved in the project.
}
