% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{tar_prep_hash}
\alias{tar_prep_hash}
\title{Generate a set of targets for fetching progress hash}
\usage{
tar_prep_hash(contents, templates = setup_templates())
}
\arguments{
\item{contents}{The contents structure used as the configuration of data
fetching.}

\item{templates}{The SQL template files used to fetch data. See
\code{\link[=setup_templates]{setup_templates()}} for details.}
}
\value{
A list of target objects.
}
\description{
The progress hash stores the progress of the project, which is used to check
whether the project is updated.
}
\details{
These objects are named as \verb{progress_hash_\{project_id\}} for each project.
}
