% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{tar_prep_proj}
\alias{tar_prep_proj}
\title{Generate a set of targets for preparing project-level data}
\usage{
tar_prep_proj(contents, templates = setup_templates(), check_progress = TRUE)
}
\arguments{
\item{contents}{The contents structure used as the configuration of data
fetching.}

\item{templates}{The SQL template files used to fetch data. See
\code{\link[=setup_templates]{setup_templates()}} for details.}

\item{check_progress}{Whether to check the progress hash. Set it as \code{FALSE}
if the project is finalized.}
}
\value{
A list of target objects.
}
\description{
There are mainly two types of data to be fetched, i.e., the progress hash and
the user information. The former is used to check the progress of the
project, while the latter is used to identify the users involved in the
project.
}
