% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{tar_prep_raw}
\alias{tar_prep_raw}
\title{Generate a set of targets for wrangling and pre-processing raw data}
\usage{
tar_prep_raw(
  contents,
  action_raw_data = c("parse", "preproc"),
  name_data = "raw_data",
  name_parsed = "raw_data_parsed",
  name_indices = "indices"
)
}
\arguments{
\item{contents}{The contents structure used as the configuration of data
fetching.}

\item{action_raw_data}{The action to be taken on the fetched raw data.}

\item{name_data}{The name of the raw data target.}

\item{name_parsed}{The name of the parsed data target.}

\item{name_indices}{The name of the indices target.}
}
\value{
A list of target objects.
}
\description{
This target factory is the main part of the \code{tar_prep_iquizoo} function. It
wrangles the raw data into a tidy format and calculates indices based on the
parsed data.
}
