% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarflow.iquizoo-package.R
\docType{package}
\name{tarflow.iquizoo-package}
\alias{tarflow.iquizoo}
\alias{tarflow.iquizoo-package}
\title{tarflow.iquizoo: Setup "targets" Workflows for "iquizoo" Data Processing}
\description{
For "iquizoo" data processing, there is already a package called "preproc.iquizoo", but eventually the use of it is relied on a workflow. This package is used to build such workflows based on tools provided by "targets" package which mimics the logic of "make", automating the building processes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://psychelzh.github.io/tarflow.iquizoo/}
  \item \url{https://github.com/psychelzh/tarflow.iquizoo}
  \item Report bugs at \url{https://github.com/psychelzh/tarflow.iquizoo/issues}
}

}
\author{
\strong{Maintainer}: Liang Zhang \email{psychelzh@outlook.com} (\href{https://orcid.org/0000-0001-9041-1150}{ORCID})

Other contributors:
\itemize{
  \item Yujian Dai [contributor]
}

}
\keyword{internal}
