% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{use_targets_template}
\alias{use_targets_template}
\title{Create standard data fetching targets pipeline script}
\usage{
use_targets_template()
}
\value{
NULL (invisible). This function is called for its side effects.
}
\description{
This function creates a standard data fetching targets pipeline script
for you to fill in.
}
