% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rawdata.R
\name{wrangle_data}
\alias{wrangle_data}
\title{Wrangle Raw Data}
\usage{
wrangle_data(data, .key = ".id")
}
\arguments{
\item{data}{The raw data.}

\item{.key}{The key name used to match meta data. Should be a scalar
character. Default is \code{".id"}, which is appropriate for most cases.}
}
\value{
A \link[dm:dm]{dm} object containing two tables: \code{meta} and \code{data}.
}
\description{
Data wrangling is the first step for data analysis.
}
\details{
These steps are performed in order:
\enumerate{
\item Data clean. Remove observations with invalid or empty (i.e., \code{"[]"} or
\code{"{}"}) json string data and then remove duplicates from data. If this step
produces data with no observation, following steps are skipped and \code{NULL}
is returned.
\item Data parse. Parse data stored in json string, convert the names to lower
case and stack the parsed data. Stacking have better performances than
\link[=group_nest]{group_nest}ing.
}
}
