% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices.R
\name{calc_indices}
\alias{calc_indices}
\title{Calculate indices for a single game}
\usage{
calc_indices(data, prep_fun)
}
\arguments{
\item{data}{The raw data.}

\item{prep_fun}{The name (symbol) of the calculation function}
}
\value{
A \code{tibble} with the calculated indices.
}
\description{
Use the given \code{prep_fun} to calculate indices for a single game. In addition,
a variable named \code{occasion} is added to track the times of participation.
}
\author{
Liang Zhang
}
