% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices.R
\name{calc_indices}
\alias{calc_indices}
\title{Calculate Indices}
\usage{
calc_indices(data, prep_fun, name_data = "game_data")
}
\arguments{
\item{data}{The raw data.}

\item{prep_fun}{The name (symbol) of the calculation function}

\item{name_data}{The column name in the \code{data} that stores original data. It
is typically of a vector containing \code{JSON} string.}
}
\value{
A \code{tibble} with the calculated indices.
}
\description{
Get raw data prepared and calculate indices.
}
\details{
These steps are performed in order:
\enumerate{
\item Remove observations with invalid or empty (i.e., \code{"[]"} or \code{"{}"}) json
string data. If this step produces data with no observation, following
steps are skipped and \code{NULL} is returned.
\item Parse data stored in json string, convert the names to lower case and
stack the parsed data. Stacking have better performances than
\link[dplyr:group_nest]{group_nest}ing.
\item Call \code{\link[dataproc.iquizoo:preproc_data]{dataproc.iquizoo::preproc_data()}} to pre-process on the parsed
data, and the results are stacked into a long format. The longer format is
used because each observation has its own creating time information.
}
}
\author{
Liang Zhang
}
