% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check_source}
\alias{check_source}
\title{Check if the database based on the given data source is ready}
\usage{
check_source(group = getOption("tarflow.group"))
}
\arguments{
\item{group}{Section identifier in the \code{default.file}. See
\code{\link[RMariaDB:dbConnect-MariaDBDriver-method]{RMariaDB::MariaDB()}} for more information.}
}
\value{
TRUE if the database is ready, FALSE otherwise.
}
\description{
Check if the database based on the given data source is ready
}
