% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_where.R
\name{compose_where}
\alias{compose_where}
\alias{compose_where.default}
\alias{compose_where.NULL}
\alias{compose_where.character}
\alias{compose_where.list}
\alias{compose_where.data.frame}
\title{Compose where clause for database query}
\usage{
compose_where(config_where, add_keyword = NULL, ...)

\method{compose_where}{default}(config_where, add_keyword = NULL, ...)

\method{compose_where}{`NULL`}(config_where, add_keyword = NULL, ...)

\method{compose_where}{character}(config_where, add_keyword = NULL)

\method{compose_where}{list}(config_where, add_keyword = NULL)

\method{compose_where}{data.frame}(config_where, add_keyword = NULL)
}
\arguments{
\item{config_where}{Configuration of "where-clause".}

\item{add_keyword}{Logical value indicates whether to add \code{"WHERE"} keyword
or not before the result. Default is \code{NULL}, which means \code{TRUE} with an
exception when \code{config_where} is \code{NULL} (it will error when setting as
\code{TRUE} in such circumstance).}

\item{...}{For extension use and must be empty.}
}
\value{
A "where-clause" character which is directly used in database query.
}
\description{
Here we just compose "where-clause" according to tables in the database.
}
\details{
In reality, this is mostly combined with a \code{yaml} configurations whose
template lives in this package, which forms an input of \code{list} data type. But
it can also handle situations with \code{character} and \code{data.frame} input.
}
\author{
Liang Zhang
}
