% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\alias{fetch_single_game}
\title{Fetch datasets from given database}
\usage{
fetch(query_file, config_where = NULL, dsn = "iquizoo-v3")

fetch_single_game(query_file, config_where = NULL, game_id, dsn = "iquizoo-v3")
}
\arguments{
\item{query_file}{File name of \code{sql} query}

\item{config_where}{Configuration of "where-clause" of the \code{sql} query. Can
be a \code{list} (mostly from the \code{config.yml} file) or \code{data.frame}.}

\item{dsn}{Data source name of an \code{odbc} database connector.}

\item{game_id}{The identifier of the game to fetch datasets from.}
}
\value{
A \code{tibble} of original data
}
\description{
Based on a template query file and "where clause" configuration, datasets are
extracted from a given database.
}
\section{Functions}{
\itemize{
\item \code{fetch}: Default usage of fetch.

\item \code{fetch_single_game}: A special case to fetch datasets from a single game.
}}

\author{
Liang Zhang
}
