% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\alias{fetch_single_game}
\title{Fetch datasets from given database}
\usage{
fetch(query_file, config_where = NULL, dsn = "iquizoo-v3", encoding = "utf-8")

fetch_single_game(query_file, game_id, config_where = NULL, ...)
}
\arguments{
\item{query_file}{File name of \code{sql} query}

\item{config_where}{Configuration of "where-clause" of the \code{sql} query. Can
be a \code{list} (mostly from the \code{config.yml} file) or \code{data.frame}.}

\item{dsn}{The data source name of an \strong{ODBC} database connector. See
\code{\link[odbc:dbConnect-OdbcDriver-method]{odbc::dbConnect()}} for more information.}

\item{encoding}{Encoding to be assumed for input strings. Default to "UTF-8".}

\item{game_id}{The identifier of the game to fetch datasets from.}

\item{...}{Other arguments passed to \code{\link[=fetch]{fetch()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} of the fetched data.
}
\description{
Based on a template query file and "where clause" configuration, datasets are
extracted from a given database.
}
\section{Functions}{
\itemize{
\item \code{fetch}: Default usage of fetch.

\item \code{fetch_single_game}: A special case to fetch datasets from a single game.
}}

\author{
Liang Zhang
}
