% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fetch_data.R
\name{fetch_config_tbl}
\alias{fetch_config_tbl}
\title{Fetch configuration table from iQuizoo database}
\usage{
fetch_config_tbl(course_name, course_period, ...)
}
\arguments{
\item{course_name}{The course name.}

\item{course_period}{The course period.}

\item{...}{Further arguments passed to \code{\link[=fetch_parameterized]{fetch_parameterized()}}.}
}
\value{
A \link{data.frame} contains the fetched data.
}
\description{
Fetch configuration table from iQuizoo database
}
