% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch data from iQuizoo database}
\usage{
fetch_data(
  query,
  project_id,
  game_id,
  course_date,
  ...,
  what = c("raw_data", "scores")
)
}
\arguments{
\item{query}{A parameterized SQL query. Note the query should also contain
a \code{glue} expression to inject the table name, i.e., \code{"{ table_name }"}.}

\item{project_id}{The project id to be bound to the query.}

\item{game_id}{The game id to be bound to the query.}

\item{course_date}{The course date. This parameter is used to determine the
table name, not to be bound to the query.}

\item{...}{Further arguments passed to \code{\link[=fetch_parameterized]{fetch_parameterized()}}.}

\item{what}{What to fetch. Can be either "raw_data" or "scores".}
}
\value{
A \link{data.frame} contains the fetched data.
}
\description{
Fetch data from iQuizoo database
}
