% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{fetch_iquizoo}
\alias{fetch_iquizoo}
\alias{fetch_iquizoo_mem}
\title{Fetch result of query from iQuizoo database}
\usage{
fetch_iquizoo(query, ..., params = NULL, source = setup_source())

fetch_iquizoo_mem(query, ..., params = NULL, source = setup_source())
}
\arguments{
\item{query}{A character string containing SQL.}

\item{...}{Further arguments passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{params}{The parameters to be bound to the query. Default to \code{NULL}, see
\code{\link[DBI:dbGetQuery]{DBI::dbGetQuery()}} for more details.}

\item{source}{The data source from which data is fetched. See
\code{\link[=setup_source]{setup_source()}} for details.}
}
\value{
A \link{data.frame} contains the fetched data.
}
\description{
Fetch result of query from iQuizoo database
}
\section{Functions}{
\itemize{
\item \code{fetch_iquizoo_mem()}: The same as \code{\link[=fetch_iquizoo]{fetch_iquizoo()}} except that the
result is cached. The cache is stored in disk by default, but can be
changed by setting the environment variable \code{TARFLOW_CACHE} to \code{"memory"}.

}}
