% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{fetch_iquizoo}
\alias{fetch_iquizoo}
\title{Fetch result of query from iQuizoo database}
\usage{
fetch_iquizoo(query, ..., params = NULL, group = getOption("tarflow.group"))
}
\arguments{
\item{query}{A character string containing SQL.}

\item{...}{Further arguments passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{params}{The parameters to be bound to the query. Default to \code{NULL}, see
\code{\link[DBI:dbGetQuery]{DBI::dbGetQuery()}} for more details.}

\item{group}{Section identifier in the \code{default.file}. See
\code{\link[RMariaDB:dbConnect-MariaDBDriver-method]{RMariaDB::MariaDB()}} for more information.}
}
\value{
A \link{data.frame} contains the fetched data.
}
\description{
Fetch result of query from iQuizoo database
}
\seealso{
\code{\link[=fetch_iquizoo_mem]{fetch_iquizoo_mem()}} for a memoised version of this function.
}
