% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{fetch_parameterized}
\alias{fetch_parameterized}
\title{Fetch data from iQuizoo database based on a parameterized query}
\usage{
fetch_parameterized(
  query,
  params,
  ...,
  dsn = NULL,
  groups = NULL,
  drv = getOption("tarflow.driver")
)
}
\arguments{
\item{query}{The query to be executed.}

\item{params}{The parameters to be bound to the query.}

\item{...}{Further arguments passed to \link[DBI:dbConnect]{DBI::dbConnect()}.}

\item{dsn}{The data source name of an \strong{ODBC} database connector. See
\code{\link[odbc:dbConnect-OdbcDriver-method]{odbc::dbConnect()}} for more information. Used when \code{drv} is set as
\code{\link[odbc:odbc]{odbc::odbc()}}.}

\item{groups}{Section identifier in the \code{default.file}. See
\code{\link[RMariaDB:dbConnect-MariaDBDriver-method]{RMariaDB::MariaDB()}} for more information. Used when \code{drv} is set as
\code{\link[RMariaDB:dbConnect-MariaDBDriver-method]{RMariaDB::MariaDB()}}.}

\item{drv}{The driver used. Set as an option of \code{"tarflow.driver"} and the
default is currently \code{odbc::odbc()}. Options are \code{\link[odbc:odbc]{odbc::odbc()}} and
\code{\link[RMariaDB:dbConnect-MariaDBDriver-method]{RMariaDB::MariaDB()}}, both of which need pre-configurations.}
}
\value{
A \link{data.frame} contains the fetched data.
}
\description{
Fetch data from iQuizoo database based on a parameterized query
}
