% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fetch_data.R
\name{fetch_preset}
\alias{fetch_preset}
\title{Fetch data from iQuizoo database with preset SQL query files}
\usage{
fetch_preset(params, what, ...)
}
\arguments{
\item{params}{The parameters used in the SQL query.}

\item{what}{The name of the preset query file to use.}

\item{...}{Further arguments passed to \code{\link[=fetch_parameterized]{fetch_parameterized()}}.}
}
\value{
A \link{data.frame} contains the fetched data.
}
\description{
Fetch data from iQuizoo database with preset SQL query files
}
