% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickup.R
\name{pickup}
\alias{pickup}
\title{Download (pick up) datasets from given database}
\usage{
pickup(query_file, config_where = NULL, dsn = "iquizoo-v3", encoding = "utf-8")
}
\arguments{
\item{query_file}{File name of \code{sql} query. Literal query is acceptable, and
to be recognized as literal query, the input must be a string containing at
least one new line.}

\item{config_where}{Configuration of "where-clause" of the \code{sql} query. Can
be a \code{list} (mostly from the \code{config.yml} file) or \code{data.frame}.}

\item{dsn}{The data source name of an \strong{ODBC} database connector. See
\code{\link[odbc:dbConnect-OdbcDriver-method]{odbc::dbConnect()}} for more information.}

\item{encoding}{Encoding to be assumed for input strings. Default to "UTF-8".}
}
\value{
A \link[tibble:tibble-package]{tibble} of the downloaded data.
}
\description{
Based on a template query file and "where clause" configuration, datasets are
extracted from a given database.
}
\section{Functions}{
\itemize{
\item \code{pickup}: Default usage of pickup.
}}

\author{
Liang Zhang
}
