% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fetch_data.R
\encoding{UTF-8}
\name{prepare_fetch_data}
\alias{prepare_fetch_data}
\title{Prepare targets pipeline for fetching data}
\usage{
prepare_fetch_data(
  tbl_params,
  ...,
  what = c("all", "raw_data", "scores"),
  cache_dir = "~/.cache.tarflow",
  cache_age = Inf
)
}
\arguments{
\item{tbl_params}{A \link{data.frame} contains the parameters to be bound to the
query. For now, only \code{course_name} and \code{course_period} are supported. Each
row is a set of parameters. See details for more information.}

\item{...}{For future usage. Should be empty.}

\item{what}{What to fetch. Can be "all", "raw_data" or "scores".}

\item{cache_dir}{The directory to store cache. Defaults to
\verb{~/.cache.tarflow}.}

\item{cache_age}{The maximum age of cache in seconds. Defaults to \code{Inf}.}
}
\value{
A list of \link[targets:tar_target]{targets}.
}
\description{
Prepare targets pipeline for fetching data
}
\details{
The \code{course_period} in \code{tble_params} could be numeric or character values.
See the following table for reference:\tabular{ll}{
   code \tab name \cr
   0 \tab 未指定 \cr
   1 \tab 学前 \cr
   2 \tab 小学低段 \cr
   3 \tab 小学中段 \cr
   4 \tab 小学高段 \cr
   5 \tab 小学 \cr
   6 \tab 初中 \cr
   7 \tab 高中 \cr
}
}
