% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{prepare_fetch_data}
\alias{prepare_fetch_data}
\title{Prepare targets based on parameters}
\usage{
prepare_fetch_data(
  params,
  ...,
  contents = NULL,
  what = c("all", "raw_data", "scores"),
  action_raw_data = c("all", "parse", "none"),
  templates = setup_templates(),
  check_progress = TRUE
)
}
\arguments{
\item{params}{A \link{data.frame} or \link{list} contains the parameters to be bound to
the query. Default templates require specifying \code{organization_name} and
\code{project_name}, in that order. If \code{contents} template is specified without
any parameters, set it as empty vector or \code{NULL}. If \code{contents} argument is
specified, this argument is omitted.}

\item{...}{For future usage. Should be empty.}

\item{contents}{The contents used as the configuration of data fetching. This
is typically automatically fetched from database based on the \code{contents}
template from \code{templates}. If not \code{NULL}, it will be used directly and
ignore that specified in \code{templates}. Note \code{contents} should at least
contains \code{project_id} and \code{game_id} names.}

\item{what}{What to fetch. If set as "all", both raw data and scores will be
fetched. If set as "raw_data", only raw data will be fetched. If set as
"scores", only scores will be fetched. Further actions on the fetched raw
data can be specified by \code{action_raw_data}.}

\item{action_raw_data}{The action to be taken on the fetched raw data. If set
as "all", both wrangling and pre-processing will be done. If set as
"parse", only wrangling will be done. If set as "none", neither will be
done. If \code{what} is "scores", this argument will be ignored.}

\item{templates}{The SQL template files used to fetch data. See
\code{\link[=setup_templates]{setup_templates()}} for details.}

\item{check_progress}{Whether to check the progress hash. Set it as \code{FALSE}
if the project is finalized.}
}
\value{
A list of target objects.
}
\description{
Given parameters, this target factory prepares a set of target objects used
to fetch data from iQuizoo database, separated into static branches so that
each is for a specific project and task/game combination.
}
