% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{prepare_fetch_data}
\alias{prepare_fetch_data}
\title{Prepare targets based on parameters}
\usage{
prepare_fetch_data(
  params,
  ...,
  what = c("all", "raw_data", "scores"),
  templates = setup_templates(),
  check_progress = TRUE
)
}
\arguments{
\item{params}{A \link{data.frame} contains the parameters to be bound to the
query. For now, only \code{organization_name} and \code{project_name} are supported
and both of them should be specified. Each row is a set of parameters.}

\item{...}{For future usage. Should be empty.}

\item{what}{What to fetch. Can be "all", "raw_data" or "scores".}

\item{templates}{The SQL template files used to fetch data. See
\code{\link[=setup_templates]{setup_templates()}} for details.}

\item{check_progress}{Whether to check the progress hash. Set it as \code{FALSE}
if the project is finalized.}
}
\value{
A list of target objects.
}
\description{
Given parameters, this target factory prepares a set of target objects used
to fetch data from iQuizoo database, separated into static branches so that
each is for a specific project and task/game combination.
}
