% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fetch_data.R
\name{prepare_fetch_data}
\alias{prepare_fetch_data}
\title{Prepare targets pipeline for fetching data}
\usage{
prepare_fetch_data(
  tbl_params,
  ...,
  what = c("raw_data", "scores"),
  cache_dir = "~/.cache.tarflow",
  cache_age = Inf
)
}
\arguments{
\item{tbl_params}{A \link{data.frame} contains the parameters to be bound to the
query. For now, only \code{course_name} and \code{course_period} are supported. Each
row is a set of parameters.}

\item{...}{For future usage. Should be empty.}

\item{what}{What to fetch. Can be either "raw_data" or "scores".}

\item{cache_dir}{The directory to store cache. Defaults to
\verb{~/.cache.tarflow}.}

\item{cache_age}{The maximum age of cache in seconds. Defaults to \code{Inf}.}
}
\value{
A list of \link[targets:tar_target]{targets}.
}
\description{
Prepare targets pipeline for fetching data
}
