% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fetch_data.R
\name{prepare_fetch_data}
\alias{prepare_fetch_data}
\title{Prepare targets pipeline for fetching data}
\usage{
prepare_fetch_data(tbl_params, ..., what = c("all", "raw_data", "scores"))
}
\arguments{
\item{tbl_params}{A \link{data.frame} contains the parameters to be bound to the
query. For now, only \code{course_name} and \code{course_period} are supported. Each
row is a set of parameters. See details for more information.}

\item{...}{For future usage. Should be empty.}

\item{what}{What to fetch. Can be "all", "raw_data" or "scores".}
}
\value{
A list of \link[targets:tar_target]{targets}.
}
\description{
Prepare targets pipeline for fetching data
}
\details{
The \code{course_period} in \code{tble_params} could be numeric or character values.
Call \code{tarflow.iquizoo:::name_course_periods} to see all the possible values.
You could input numeric index, for example, \code{1} means the first in the
\code{tarflow.iquizoo:::name_course_periods}. Use \code{0} or \code{NA} if you want to refer
to older classes with no course periods.
}
