% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_data.R
\name{preproc_data}
\alias{preproc_data}
\title{Feed Raw Data to Pre-processing}
\usage{
preproc_data(dm, ...)
}
\arguments{
\item{dm}{A \link[dm:dm]{dm} object typically from \code{\link[=wrangle_data]{wrangle_data()}} contains
two tables: \code{meta} and \code{data}.}

\item{...}{Parameters passed to \code{\link[preproc.iquizoo:preproc]{preproc.iquizoo::preproc()}}.}
}
\value{
A \link[dm:dm]{dm} object containing three tables: \code{meta}, \code{data} and
\code{indices}.
}
\description{
Integrate \link[dm:dm]{dm} object returned by \code{\link[=wrangle_data]{wrangle_data()}} with
\code{\link[preproc.iquizoo:preproc]{preproc.iquizoo::preproc()}} API.
}
