% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{preproc_data}
\alias{preproc_data}
\title{Feed Raw Data to Pre-processing}
\usage{
preproc_data(
  data,
  fn,
  ...,
  name_raw_parsed = "raw_parsed",
  out_name_index = "index_name",
  out_name_score = "score"
)
}
\arguments{
\item{data}{A \link{data.frame} contains raw data.}

\item{fn}{This can be a function or formula. See \code{\link[rlang:as_function]{rlang::as_function()}} for
more details.}

\item{...}{Additional arguments passed to \code{fn}.}

\item{name_raw_parsed}{The column name in which stores user's raw data in
format of a list of \link[=data.frame]{data.frame}s.}

\item{out_name_index}{The column name used in output storing the name of each
calculated index.}

\item{out_name_score}{The column name used in output storing the value of
each calculated index.}
}
\value{
A \link{data.frame} contains the calculated indices. The index names are
stored in the column of \code{out_name_index}, and index values are stored in
the column of \code{out_name_score}.
}
\description{
Calculate indices using data typically returned by \code{\link[=wrangle_data]{wrangle_data()}}.
}
\details{
Observations with empty raw data (empty vector, e.g. \code{NULL}, in
\code{name_raw_parsed} column) are removed before calculating indices. If no
observations left after removing, a warning is signaled and \code{NULL} is
returned.
}
