% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_data.R
\name{preproc_data}
\alias{preproc_data}
\title{Feed Raw Data to Pre-processing}
\usage{
preproc_data(data, fn, name_raw_parsed = "raw_parsed", ...)
}
\arguments{
\item{data}{A \link[tibble:tibble-package]{tibble} contains raw data.}

\item{fn}{This can be a function or formula. See \code{\link[rlang:as_function]{rlang::as_function()}} for
more details.}

\item{name_raw_parsed}{The column name in which stores user's raw data in
format of a list of \link[=data.frame]{data.frame}s.}

\item{...}{Additional arguments passed to \code{fn}.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains the calculated indices.
}
\description{
Integrate data returned by \code{\link[=wrangle_data]{wrangle_data()}} with functions in preproc.iquizoo
package.
}
\details{
The calculated indices are stored in a \link[=data.frame]{data.frame} of
\link[tidyr:pivot_longer]{longer} format with two columns named \code{"index_name"}
and \code{"score"} respectively and are nested into a
\link[tibble:tibble-package]{tibble} object. The longer format is used to
facilitate combining.
}
