% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_games.R
\name{search_games}
\alias{search_games}
\alias{search_games_mem}
\title{Search and cache games used in analysis}
\usage{
search_games(config_where, known_only = TRUE, query_file = NULL)

search_games_mem(file_cache = "~/.cache.tarflow", ...)
}
\arguments{
\item{config_where}{Configuration of "where-clause".}

\item{known_only}{Logical value indicates whether to use games in
\code{\link[data.iquizoo:game_info]{game_info}} only (default) or not.}

\item{query_file}{An optional argument specifying the file storing query of
games. If leave as \code{NULL}, default to "sql/games.tmpl.sql", which is
created by rmarkdown template.}

\item{file_cache}{The file used to store caches of results which is passed to
\code{\link[cachem:cache_disk]{cachem::cache_disk()}}.}

\item{...}{Arguments passed to \code{\link[=search_games]{search_games()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains all the games to be
analyzed and its related information.
}
\description{
When separate analysis into branches, the games to be analyzed should be
known in advance. This function should be used with memoise package to reduce
executing time.
}
\details{
The fetched will be joined with \code{\link[data.iquizoo:game_info]{game_info}} data
from data.iquizoo package to expose the pre-processing function name.
}
\section{Functions}{
\itemize{
\item \code{search_games_mem}: Cached version using
\link[memoise:memoise]{memoise()}.
}}

