% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_games.R
\name{search_games}
\alias{search_games}
\alias{search_games_mem}
\title{Search and cache games used in analysis}
\usage{
search_games(config_where)
# cached version using `memoise::memoise()`
search_games_mem(config_where)
}
\arguments{
\item{config_where}{Configuration of "where-clause".}

\item{known_only}{Logical value indicates whether to use games in
\code{\link[dataproc.iquizoo:game_info]{game_info}} only (default) or not.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains all the games to be
analyzed and its related information.
}
\description{
When separate analysis into branches, the games to be analyzed should be
known in advance. This function should be used with memoise package to reduce
executing time.
}
\details{
The fetched will be joined with \code{\link[dataproc.iquizoo:game_info]{game_info}}
data from dataproc.iquizoo package to expose the pre-processing function
name.
}
