% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{setup_option_file}
\alias{setup_option_file}
\title{Setup MySQL database connection option file}
\usage{
setup_option_file(path = NULL, overwrite = FALSE)
}
\arguments{
\item{path}{The path to the option file. Default location is operating system
dependent. On Windows, it is \verb{C:/my.cnf}. On other systems, it is
\verb{~/.my.cnf}.}

\item{overwrite}{Whether to overwrite the existing option file.}
}
\value{
NULL (invisible).
}
\description{
This function will create a MySQL option file at the given path. To ensure it
works, set these environment variables before calling this function:
\itemize{
\item \code{MYSQL_HOST}: The host name of the MySQL server.
\item \code{MYSQL_USER}: The user name of the MySQL server.
\item \code{MYSQL_PASSWORD}: The password of the MySQL server.
}
}
