% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{setup_templates}
\alias{setup_templates}
\title{Set up templates used to fetch data}
\usage{
setup_templates(
  contents = NULL,
  users = NULL,
  raw_data = NULL,
  scores = NULL,
  progress_hash = NULL
)
}
\arguments{
\item{contents}{The SQL template file used to fetch contents. At least
\code{project_id}, \code{game_id} and \code{course_date} should be included in the
contents.}

\item{users}{The SQL template file used to fetch users.}

\item{raw_data}{The SQL template file used to fetch raw data. See
\code{\link[=fetch_data]{fetch_data()}} for details. Usually you don't need to change this.}

\item{scores}{The SQL template file used to fetch scores. See \code{\link[=fetch_data]{fetch_data()}}
for details. Usually you don't need to change this.}

\item{progress_hash}{The SQL template file used to fetch progress hash.
Usually you don't need to change this.}
}
\value{
A S3 object of class \code{tarflow.template} with the options.
}
\description{
If you want to extract data based on your own parameters, you should use this
function to set up your own SQL templates. Note that the SQL queries should
be parameterized.
}
