% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fetch_data.R
\name{setup_templates}
\alias{setup_templates}
\title{Set up templates used to fetch data}
\usage{
setup_templates(contents = NULL, users = NULL, progress_hash = NULL)
}
\arguments{
\item{contents}{The SQL template file used to fetch contents. At least
\code{project_id}, \code{game_id} and \code{course_date} should be included in the
contents. See \code{\link[=fetch_data]{fetch_data()}} for details.}

\item{users}{The SQL template file used to fetch users. See \code{\link[=fetch_batch]{fetch_batch()}}
for details.}

\item{progress_hash}{The SQL template file used to fetch progress hash. See
\code{\link[=fetch_batch]{fetch_batch()}} for details. Usually you don't need to change this.}
}
\value{
A S3 object of class \code{tarflow.template} with the options.
}
\description{
Set up templates used to fetch data
}
