% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{tar_prep_iquizoo}
\alias{tar_prep_iquizoo}
\title{Generate a set of targets for pre-processing of iQuizoo data}
\usage{
tar_prep_iquizoo(
  params,
  ...,
  contents = NULL,
  what = c("raw_data", "scores"),
  action_raw_data = c("all", "parse", "none"),
  templates = setup_templates(),
  check_progress = TRUE
)
}
\arguments{
\item{params}{A \link{data.frame} or \link{list} contains the parameters to be bound to
the query. Default templates require specifying \code{organization_name} and
\code{project_name}, in that order. If \code{contents} template is specified without
any parameters, set it as empty vector or \code{NULL}. If \code{contents} argument is
specified, this argument is omitted.}

\item{...}{For future usage. Should be empty.}

\item{contents}{The contents structure used as the configuration of data
fetching. It is typically automatically fetched from database based on the
\code{contents} template in \code{templates}. If not \code{NULL}, it will be used directly
and ignore that specified in \code{templates}. Note \code{contents} should at least
contains \code{project_id} and \code{game_id} names.}

\item{what}{What to fetch. There are basically two types of data, i.e., raw
data and scores. The former is the logged raw data for each trial of the
tasks/games, while the latter is the scores calculated by iQuizoo server.
If set as "all", both raw data and scores will be fetched. Further actions
on the fetched raw data can be specified by \code{action_raw_data}.}

\item{action_raw_data}{The action to be taken on the fetched raw data. There
are two consecutive actions, i.e., wrangling and pre-processing. The former
will parse the raw data into a tidy format, while the latter will calculate
indices based on the parsed data. If set as "all", both wrangling and
pre-processing will be done. If set as "parse", only wrangling will be
done. If set as "none", neither will be done. If \code{what} is "scores", this
argument will be ignored.}

\item{templates}{The SQL template files used to fetch data. See
\code{\link[=setup_templates]{setup_templates()}} for details.}

\item{check_progress}{Whether to check the progress hash. Set it as \code{FALSE}
if the project is finalized.}
}
\value{
A list of target objects.
}
\description{
This target factory prepares a set of target objects used to fetch data from
iQuizoo database, separated into static branches so that each is for a
specific project and task/game combination. Further pre-processing on the
fetched data can also be added if requested.
}
