% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TarScript.R
\name{TarScript}
\alias{TarScript}
\title{TarScript class}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Stores all the required commands used
for pipeline building. See \code{\link[targets:tar_script]{targets::tar_script()}} for more details of a
target script.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TarScript$new()}}
\item \href{#method-build}{\code{TarScript$build()}}
\item \href{#method-update}{\code{TarScript$update()}}
\item \href{#method-print}{\code{TarScript$print()}}
\item \href{#method-deparse_script}{\code{TarScript$deparse_script()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \link{TarScript} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TarScript$new(
  package = c("targets", "tarchetypes"),
  global = NULL,
  option = NULL,
  targets = NULL,
  pipeline = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{package}}{Character vector. The package need to be included in the
script. Usually only "targets" and "tarchetypes". Set it \code{NULL} if not
needed. And repetitive packages are removed.}

\item{\code{global}}{A \code{character} vector. Global variables or functions used in
the script. For example, there might be a \code{source()} call to prepare
functions. \code{character} used for better formatting. No \code{unique()} is
applied.}

\item{\code{option}}{A \code{list}. Will be arguments of \code{tar_option_set()}. Set it
\code{NULL} if not needed.}

\item{\code{targets}}{A \code{character} vector. Defined targets out of the pipeline
\code{list}. \code{character} used for better formatting. No \code{unique()} is
applied.}

\item{\code{pipeline}}{A \code{list} of expressions, each of which is a \code{call} to
\code{tar_target()} or its related. This cannot be \code{NULL} when build.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build"></a>}}
\if{latex}{\out{\hypertarget{method-build}{}}}
\subsection{Method \code{build()}}{
Build the targets script. This will update the file "_targets.R" from
current project by default.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TarScript$build(path = NULL, ..., styler = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{A \link[base:connections]{connection}, or a character string
naming the file to print the script. Set to "" to print to the standard
output. The default is "_targets.R".}

\item{\code{...}}{For future expansion use and must be empty.}

\item{\code{styler}}{A logical indicating if styler should be called to make the
generated file nicer to read. Default is \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Update a part of the script.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TarScript$update(step, codes, append = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step}}{A character of the updating step name.}

\item{\code{codes}}{The codes to update.}

\item{\code{append}}{A logical value indicating if the \code{codes} are appended to
the old \code{step} (\code{TRUE}) or replaced (\code{FALSE}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print object more informative.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TarScript$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{For future expansion use and must be empty.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deparse_script"></a>}}
\if{latex}{\out{\hypertarget{method-deparse_script}{}}}
\subsection{Method \code{deparse_script()}}{
Deparse the whole script to a character vector. Each element of it is a
code line. Note this method does not return the object itself, so it
cannot be chained.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TarScript$deparse_script()}\if{html}{\out{</div>}}
}

}
}
