% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_targets.R
\name{use_targets}
\alias{use_targets}
\title{Add specific targets to workflow}
\usage{
use_targets(
  schema = c("indices", "scores", "original"),
  separate = NULL,
  ignore_tar = TRUE
)
}
\arguments{
\item{schema}{The name of your jobs. 'indices' means calculate game indices
from rawdata, mostly for scholar usage; 'scores' means extracting the
scores directly from online database, mostly for norm calculation and
report generation; 'original' fetches raw data only.}

\item{separate}{Separate main targets into branches or not. For \code{"indices"}
and \code{"original"} schema, the default is set to \code{TRUE}. For \code{"scores"}
schema, the default is set to \code{FALSE}. Note, when setting to \code{TRUE},
package "dataproc.iquizoo" is required and its version should be \code{"0.2.6"}
or higher.}

\item{ignore_tar}{Should internal data generated by "targets" making be
ignored or not? Default to \code{TRUE}.}
}
\description{
Add some pre-defined targets to your workflow.
}
\author{
Liang Zhang
}
