% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_targets.R
\name{use_targets}
\alias{use_targets}
\title{Add specific targets to workflow}
\usage{
use_targets(schema = c("indices", "scores"))
}
\arguments{
\item{schema}{The name of your jobs. 'indices' means calculate game idices
from rawdata, mostly for scholar usage; and 'scores' means extracting the
scores directly from online database, mostly for norm calculation and
report generation.}
}
\description{
Add some pre-defined targets to your workflow.
}
\author{
Liang Zhang
}
