% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{wrangle_data}
\alias{wrangle_data}
\title{Wrangle Raw Data}
\usage{
wrangle_data(data, name_raw_json = "game_data", name_raw_parsed = "raw_parsed")
}
\arguments{
\item{data}{The raw data.}

\item{name_raw_json}{The column name in which stores user's raw data in
format of json string.}

\item{name_raw_parsed}{The name used to store parsed data.}
}
\value{
A \link{data.frame} contains the parsed data.
}
\description{
Parse raw json string data as \code{\link[=data.frame]{data.frame()}} and store them in a list column.
}
