% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data.R
\name{wrangle_data}
\alias{wrangle_data}
\title{Wrangle Raw Data}
\usage{
wrangle_data(data, name_raw_json = "game_data")
}
\arguments{
\item{data}{The raw data.}

\item{name_raw_json}{The column name in which stores user's raw data in
format of json string.}
}
\value{

}
\description{
Data wrangling is the first step for data analysis.
}
